refactor: 유저정보 요청 메소드 이동 from AuthController -> to MemberController
