+++ src/main/java/app/iris/missingyou/controller/AuthController.java
@@ -1,18 +1,14 @@
 package app.iris.missingyou.controller;
 
-import app.iris.missingyou.dto.MemberInfoDto;
 import app.iris.missingyou.dto.TokenDto;
 import app.iris.missingyou.entity.Member;
 import app.iris.missingyou.entity.Platform;
-import app.iris.missingyou.security.CustomUserDetails;
 import app.iris.missingyou.security.GoogleUser;
 import app.iris.missingyou.security.OAuth2Service;
 import app.iris.missingyou.service.MemberService;
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -39,16 +35,6 @@ public class AuthController {
     @GetMapping(value = "/auth/refresh")
     public ResponseEntity<TokenDto> reIssueToken(@RequestHeader(required = true, name = "refreshToken") String refreshToken) {
         TokenDto dto = memberService.reIssueToken(refreshToken);
-        return ResponseEntity.ok().body(dto);
-    }
-
-    @Operation(summary = "유저 정보 요청", security = { @SecurityRequirement(name = "bearerAuth")})
-    @GetMapping(value = "/member")
-    public ResponseEntity<MemberInfoDto> getUserInfo() {
-        CustomUserDetails principal = (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        MemberInfoDto dto = memberService.get(principal);
-
         return ResponseEntity.ok().body(dto);
     }
 }
